support VK_CHANGE_INPUTMODE @ kext

+++ src/core/kext/Client.cpp
@@ -114,7 +114,7 @@ namespace org_pqrs_KeyRemap4MacBook {
   }
 
   int
-  KeyRemap4MacBook_client::sendmsg(KeyRemap4MacBook_bridge::RequestType type, void* request, size_t requestsize, void* reply, size_t replysize)
+  KeyRemap4MacBook_client::sendmsg(KeyRemap4MacBook_bridge::RequestType type, void* request, uint32_t requestsize, void* reply, uint32_t replysize)
   {
     if (! lock_) { return EIO; }
